[CWS] extract probe monitors
